implement public interface of this node first
