+++ perception/autoware_euclidean_cluster_object_detector/src/euclidean_cluster_object_detector_node.hpp
@@ -15,60 +15,4 @@
 #ifndef EUCLIDEAN_CLUSTER_OBJECT_DETECTOR_NODE_HPP_
 #define EUCLIDEAN_CLUSTER_OBJECT_DETECTOR_NODE_HPP_
 
-#include "parameters.hpp"
-
-#include <tl_expected/expected.hpp>
-
-#include <pcl/point_cloud.h>
-#include <pcl/point_types.h>
-
-#include <string>
-#include <vector>
-
-namespace autoware::euclidean_cluster
-{
-
-class EuclideanClusterObjectDetector
-{
-public:
-  explicit EuclideanClusterObjectDetector(const EuclideanClusterParams & param);
-
-  // Single public interface for everything
-  tl::expected<std::vector<pcl::PointCloud<pcl::PointXYZ>>, std::string> cluster(
-    const pcl::PointCloud<pcl::PointXYZ>::ConstPtr & input_cloud) const;
-
-private:
-  EuclideanClusterParams param_;
-
-  /**
-   * @brief Helper func to perform standard Euclidean clustering on a point cloud.
-   *        Basically this is the brute-force approach without downsampling, a.k.a slower but
-   * accurate (I guess?). This algorithm loads point cloud into 3D KD-Tree for O(logN) nearest
-   * neighbor search, then performs clustering based on distance tolerance and min/max cluster size.
-   *
-   * @param input_cloud Input point cloud to cluster.
-   *
-   * @return tl::expected<std::vector<pcl::PointCloud<pcl::PointXYZ>>, std::string> A vector of
-   * point clouds, each representing a cluster.
-   */
-  [[nodiscard]] tl::expected<std::vector<pcl::PointCloud<pcl::PointXYZ>>, std::string>
-  cluster_standard(const pcl::PointCloud<pcl::PointXYZ>::ConstPtr & input_cloud) const;
-
-  /**
-   * @brief Helper func to perform voxel grid downsampling on a point cloud, then perform Euclidean
-   * clustering. This is a faster but less accurate approach, as downsampling may remove points
-   * important for clustering. Still, it's useful for large point clouds where speed is more
-   * important than accuracy.
-   *
-   * @param input_cloud Input point cloud to cluster.
-   *
-   * @return tl::expected<std::vector<pcl::PointCloud<pcl::PointXYZ>>, std::string> A vector of
-   * point clouds, each representing a cluster.
-   */
-  [[nodiscard]] tl::expected<std::vector<pcl::PointCloud<pcl::PointXYZ>>, std::string>
-  cluster_voxel_grid(const pcl::PointCloud<pcl::PointXYZ>::ConstPtr & input_cloud) const;
-};
-
-}  // namespace autoware::euclidean_cluster
-
 #endif  // EUCLIDEAN_CLUSTER_OBJECT_DETECTOR_NODE_HPP_
